requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.1
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cmake==4.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.1.4
debugpy==1.8.15
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
easydict==1.13
executing==2.2.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
Faker==40.8.0
fastjsonschema==2.21.1
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.1
fqdn==1.5.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
googletrans==4.0.0rc1
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
insightface==0.7.3
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.7.0
jaxlib==0.7.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mediapipe==0.10.21
mistune==3.1.3
ml_dtypes==0.5.3
mpmath==1.3.0
narwhals==2.11.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.19.0
onnxruntime==1.22.1
openai==1.108.1
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathfinding==1.0.18
pillow==10.4.0
platformdirs==4.3.8
plotly==6.6.0
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pymongo==4.16.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==3.3.0
pytz==2025.1
pywin32==311
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==27.0.1
ratelim==0.1.6
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.1
setuptools==80.9.0
simsimd==6.5.3
six==1.16.0
sketchpy==0.1.7
slack_sdk==3.40.1
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
streamlit==1.51.0
streamlit-plotly-events==0.0.6
stringzilla==4.0.14
svg.path==6.3
svgpathtools==1.6.1
svgwrite==1.4.3
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.9.20
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
turtle==0.0.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.1
uri-template==1.3.0
urllib3==2.2.3
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.14
